requirements.txt
aiofiles==25.1.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
DateTime==6.0
ddgs==9.9.3
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
eventlet==0.33.3
fake-useragent==2.2.0
fastapi==0.125.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
fonttools==4.61.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==0.3.0
google-generativeai==0.8.6
google_search_results==2.4.2
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
multipledispatch==1.0.0
multitasking==0.0.12
narwhals==2.14.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
primp==0.15.0
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pykrx==1.0.51
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.13.0
python-socketio==5.10.0
pytz==2025.2
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
rsa==4.9.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.1
starlette==0.50.0
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
treys==0.1.8
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.4
wrapt==2.0.1
wsproto==1.3.2
xlrd==2.0.2
yfinance==0.2.28
zope.interface==8.1.1
